feat: create UI and merge api review in admin, user

+++ PatchSeller.DAL/Models/Review.cs
@@ -11,7 +11,7 @@ public class Review
     public string UserName { get; set; }
     public string Title { get; set; }
     public string Content { get; set; }
-    public double Overall { get; set; }
+    public int Overall { get; set; }
     public int Status { get; set; }
     public DateTime? CreatedAt { get; set; }
